% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swim_yaps.R
\name{swim_yaps}
\alias{swim_yaps}
\title{functions to get receiver and tagging metadata into session}
\usage{
swim_yaps(fish_detections, runs, rbi_min, rbi_max)
}
\arguments{
\item{fish_detections}{the data frame with the fish detections including epo and frac}

\item{runs}{number of times to refit the model}

\item{rbi_min}{the minimum random burst interval of the transmitter}

\item{rbi_max}{the maximum random burst interval of the transmitter}
}
\description{
functions to get receiver and tagging metadata into session
}
