% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{functions to get receiver and tagging metadata into session}
\arguments{
\item{hydros}{the hydrophone data frame as a data.table}

\item{detections}{a data.table of the sync tag detections}

\item{ss_data}{a data.table of the speed of sound data formatted using yaps::tempToSs}

\item{HOW_THIN}{is the number for the eps_threshold.. smaller makes a thinner distribution default is 50}

\item{keep_rate}{what proportion of sync tag detections do you want to retain, default is 1 (100\%)}

\item{timekeeper}{is the idx of the most perfectly fixed hydrophone; refer to hydros data.table for idx numbers}
}
\description{
functions to get receiver and tagging metadata into session
}
