% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaps_all.R
\name{yaps_all}
\alias{yaps_all}
\title{functions to get receiver and tagging metadata into session}
\arguments{
\item{fish_detections}{the dataframe with the fish detections including epo and frac}

\item{hydros}{the dataframe with the hydrophones and sync tags set up as per yaps}

\item{sync_model}{the sync model you generated with yaps}

\item{rbi_min}{the minimum random burst interval of the transmitter}

\item{rbi_max}{the maximum random burst interval of the transmitter}

\item{transmitter_ID}{the ID of the transmitter you wanted to position}

\item{date}{the date that you wanted to position the tag on}

\item{runs}{number of times to refit the model}
}
\description{
functions to get receiver and tagging metadata into session
}
