% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscYAPS-package.R
\docType{package}
\name{miscYAPS-package}
\alias{miscYAPS}
\alias{miscYAPS-package}
\title{miscYAPS: Miscellaneous data and helper functions for using YAPS}
\description{
Provides a bit of data and helper functions for wrapping around the key YAPS functions
}
\author{
\strong{Maintainer}: Robert Lennox \email{lennox@dal.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
